refactor: move app/constants.py to routes/
